requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
pyjwt
